[ll] core Move window system interface related parts into separate module and improve docs
